refactor: Phase 5 slice 5f — components/ folder reorg (#118)
Reorganizes the flat components/ directory into three sibling subfolders:

  components/
    views/      → AnimeDetailView, CalendarView, DownloadsView,
                  FriendsActivityView, HomeView, LibraryView, PlayerView,
                  SearchView, SettingsView, ShikimoriView (the 10 route
                  views imported by App.vue)
    shared/     → AnimeCard, CleanupModal, Sidebar (cross-view atoms)
    detail/     → AnimeDetailView's panel + episode-list children
                  (unchanged location — sibling, not nested under views/)
    player/     → PlayerView's title bar + menu children (unchanged)
    settings/   → SettingsView's tab components (unchanged)

The detail/, player/, settings/ subfolders stay at the components/ root
on purpose: they're already siblings of their parent views and keeping
them flat means the existing within-folder relative imports
(./EpisodeRow.vue, etc.) don't churn.

Mechanical change only — no behavior change, no new files, no logic
edits. PlayerView.vue still ships the inline composables wired up in
slice 5d.

Import path updates:
- App.vue's 12 component imports retargeted to ./components/views/
  or ./components/shared/.
- Each moved view file's relative imports shift one level deeper:
  '../utils' → '../../utils',
  '../stores/X' → '../../stores/X',
  '../composables/X' → '../../composables/X'.
- Cross-component imports inside views/:
  '\./AnimeCard.vue' → '../shared/AnimeCard.vue'  (SearchView,
    LibraryView, ShikimoriView, HomeView)
  '\./CleanupModal.vue' → '../shared/CleanupModal.vue'  (AnimeDetailView)
  '\./detail/X.vue' → '../detail/X.vue'  (AnimeDetailView)
  '\./player/X.vue' → '../player/X.vue'  (PlayerView)
  '\./settings/SettingsShell.vue' → '../settings/SettingsShell.vue'
    (SettingsView)
- Shared components' relative imports shift one level deeper too
  (../utils → ../../utils, ../stores → ../../stores).

DESIGN.md: updates the file-map paths for the moved components.

Quality gate: typecheck + lint + format:check + 318 tests + build +
subscription-contract grep gate all green.

Stacked on #134 (slice 5e). Last planned Phase 5 slice.
