+++ src/renderer/src/App.vue
@@ -6,18 +6,18 @@ import { usePlayerStore } from './stores/player';
 import { useSettingsStore } from './stores/settings';
 import { useShikimoriStore } from './stores/shikimori';
 import { useKeyboardShortcuts, type ShortcutAction } from './composables/keyboard-shortcuts';
-import Sidebar from './components/Sidebar.vue';
-import SearchView from './components/SearchView.vue';
-import LibraryView from './components/LibraryView.vue';
-import SettingsView from './components/SettingsView.vue';
-import DownloadsView from './components/DownloadsView.vue';
-import AnimeDetailView from './components/AnimeDetailView.vue';
-import ShikimoriView from './components/ShikimoriView.vue';
-import FriendsActivityView from './components/FriendsActivityView.vue';
-import CalendarView from './components/CalendarView.vue';
-import HomeView from './components/HomeView.vue';
-import PlayerView from './components/PlayerView.vue';
-import CleanupModal from './components/CleanupModal.vue';
+import Sidebar from './components/shared/Sidebar.vue';
+import CleanupModal from './components/shared/CleanupModal.vue';
+import SearchView from './components/views/SearchView.vue';
+import LibraryView from './components/views/LibraryView.vue';
+import SettingsView from './components/views/SettingsView.vue';
+import DownloadsView from './components/views/DownloadsView.vue';
+import AnimeDetailView from './components/views/AnimeDetailView.vue';
+import ShikimoriView from './components/views/ShikimoriView.vue';
+import FriendsActivityView from './components/views/FriendsActivityView.vue';
+import CalendarView from './components/views/CalendarView.vue';
+import HomeView from './components/views/HomeView.vue';
+import PlayerView from './components/views/PlayerView.vue';
 import { formatBytes } from './utils';
 
 const libraryStore = useLibraryStore();
