+++ src/renderer/src/components/views/FriendsActivityView.vue
@@ -1,6 +1,6 @@
 <script setup lang="ts">
 import { ref, onMounted } from 'vue';
-import { useLibraryStore } from '../stores/library';
+import { useLibraryStore } from '../../stores/library';
 
 const libraryStore = useLibraryStore();
